Revert "fix: Iterator on immutable bitmap didn't work, as it referenced container positions using a short instead of an int."
This reverts commit 97536d333691ddd0b4ebfeacbc8c61ca9395769d.

+++ src/main/java/org/roaringbitmap/buffer/ImmutableRoaringBitmap.java
@@ -1088,7 +1088,7 @@ public class ImmutableRoaringBitmap
 
       ShortIterator iter;
 
-      int pos = 0;
+      short pos = 0;
 
       int x;
